Restore YAML code view + adjust hints
